requirements.txt
django==2.2.2
djangorestframework==3.9.4